format everything and add intellij project files

+++ src/net/server/channel/handlers/PetCommandHandler.java
@@ -23,9 +23,9 @@ package net.server.channel.handlers;
 
 import client.MapleCharacter;
 import client.MapleClient;
+import client.inventory.MaplePet;
 import client.inventory.PetCommand;
 import client.inventory.PetDataFactory;
-import client.inventory.MaplePet;
 import net.AbstractMaplePacketHandler;
 import tools.MaplePacketCreator;
 import tools.Randomizer;
@@ -47,11 +47,11 @@ public final class PetCommandHandler extends AbstractMaplePacketHandler {
         slea.readInt();
         slea.readByte();
         byte command = slea.readByte();
-        PetCommand petCommand = PetDataFactory.getPetCommand(pet.getItemId(), (int) command);
+        PetCommand petCommand = PetDataFactory.getPetCommand(pet.getItemId(), command);
         if (petCommand == null) {
             return;
         }
-        
+
         if (Randomizer.nextInt(100) < petCommand.getProbability()) {
             pet.gainClosenessFullness(chr, petCommand.getIncrease(), 0, command);
             chr.getMap().broadcastMessage(MaplePacketCreator.commandResponse(chr.getId(), petIndex, false, command, false));
